Wrote a function to convert nested JSON into a Pandas DataFrame columns.
For the AWS Mechanical Turk dataset (column 4) included a cell value with JSON event information. In order for us to be able to read information from this JSON object wrote a function to flatten the JSON data into separate dataframe columns with prefix of the original column name plus suffix of the nested JSON field names.

This will allow us to filter the dataset based on this nested JSON data.